+++ src/group/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 from course.models import Course
+from user.models import Student
 
 
 
@@ -12,3 +13,12 @@ class Group(models.Model):
     
     def __str__(self):
         return f"{self.name} ({self.course.name})"
+    
+class StudentGroup(models.Model):
+    group_id = models.ForeignKey(Group, on_delete=models.CASCADE)
+    student_id = models.ForeignKey(Student, on_delete=models.CASCADE)
+    status = models.BooleanField(default=False)
+    joined_at = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return f'{self.group_id} - {self.student_id}'
